+++ NEWS.md
@@ -4,3 +4,5 @@
 
 - Add `basilisk` environment.
 - Add `spectraVariableMapping`.
+- Refactor functions to convert between R and python spectrum objects and add
+  unit tests.